+++ osquery/tables/system/linux/deb_package_helpers.h
